// src/app/model/prevision.ts
import {Session} from "./session";
import {Profil} from "./profil";

export class Prevision {
  revenus_annuel?: number;
  revenus_annuel_previsionnel?: number
  benefice_annuel_net?:number;
  benefice_annuel_net_previsionnel?: number;
  periodicite_activite?: string ;
  periodicite_activite_previsionnel?: string;
}